Fixes MPLP.map_query to return the variable states instead of probability values [fixes #1498]
